Add test to check that metadata doesn't change
