+++ interop-tests/Cargo.toml
@@ -12,10 +12,13 @@ serde = "1"
 serde_derive = "1"
 serde_json = "1"
 tuf = { version = "0.3.0-beta1", path = "../tuf" }
+walkdir = "2.3.2"
 
 [dev-dependencies]
-matches = "0.1.8"
 futures-util = { version = "0.3.1", features = [ "io" ] }
+matches = "0.1.8"
+pretty_assertions = "1"
+tempfile = "3"
 
 [[bin]]
 name = "generate"
